Create WhoLikesIt.cpp
You probably know the "like" system from Facebook and other pages. 
People can "like" blog posts, pictures or other items. We want to 
create the text that should be displayed next to such an item.

Implement the function which takes an array containing the names of people that like an item.